Ability to update checking database when withdraw

+++ ATM_Project/src/main/java/com/cs3560/atm_project/Credit.java
@@ -41,7 +41,7 @@ public class Credit extends Account{
         }
     }
     
-    public boolean updateCreditUsed(int amount) {
+    public boolean updateCreditUsed(double amount) {
         boolean result = false;
         creditUsed += amount;
         if (creditLimit > creditUsed) {
